coleccionProvincias arreglado

+++ src/main/java/modulos/servicioEstadistica/entities/ColeccionProvincia.java
@@ -22,12 +22,12 @@ public class ColeccionProvincia {
     @Column(name = "provincia_id")
     private Long provincia_id;
     @Column(name = "cantidad_hechos")
-    private Integer cantidad;
+    private Long cantidad;
 
     @Column(name = "timestamp")
     private Instant timestamp;
 
-    public ColeccionProvincia(Long coleccion_id, Long provincia_id, Integer cantidad) {
+    public ColeccionProvincia(Long coleccion_id, Long provincia_id, Long cantidad) {
         this.coleccion_id = coleccion_id;
         this.provincia_id = provincia_id;
         this.cantidad = cantidad;
